requirements.txt
astroid==2.4.2
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.4.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pylint==2.6.0
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
